Sanitise: Create new sub-package
This now separates associates and sequence assoction utilities and
defines the `SanitiseTransformation` in the sub-pacakge root.
