+++ loki/transformations/sanitise/associates.py
@@ -12,51 +12,16 @@ constructs to unify code structure and make reasoning about Fortran
 code easier.
 """
 
-from loki.batch import Transformation
-from loki.expression import Array, RangeIndex, LokiIdentityMapper
-from loki.ir import nodes as ir, FindNodes, Transformer, NestedTransformer
+from loki.expression import LokiIdentityMapper
+from loki.ir import nodes as ir, Transformer, NestedTransformer
 from loki.scope import SymbolTable
-from loki.tools import as_tuple, dict_override
-from loki.types import BasicType
+from loki.tools import dict_override
 
 
 __all__ = [
-    'SanitiseTransformation', 'resolve_associates', 'merge_associates',
-    'ResolveAssociatesTransformer', 'transform_sequence_association',
-    'transform_sequence_association_append_map'
+    'resolve_associates', 'merge_associates',
+    'ResolveAssociatesTransformer'
 ]
-
-
-class SanitiseTransformation(Transformation):
-    """
-    :any:`Transformation` object to apply several code sanitisation
-    steps when batch-processing large source trees via the :any:`Scheduler`.
-
-    Parameters
-    ----------
-    resolve_associate_mappings : bool
-        Resolve ASSOCIATE mappings in body of processed subroutines; default: True.
-    resolve_sequence_association : bool
-        Replace scalars that are passed to array arguments with array
-        ranges; default: False.
-    """
-
-    def __init__(
-            self, resolve_associate_mappings=True, resolve_sequence_association=False
-    ):
-        self.resolve_associate_mappings = resolve_associate_mappings
-        self.resolve_sequence_association = resolve_sequence_association
-
-    def transform_subroutine(self, routine, **kwargs):
-
-        # Associates at the highest level, so they don't interfere
-        # with the sections we need to do for detecting subroutine calls
-        if self.resolve_associate_mappings:
-            resolve_associates(routine)
-
-        # Transform arrays passed with scalar syntax to array syntax
-        if self.resolve_sequence_association:
-            transform_sequence_association(routine)
 
 
 def resolve_associates(routine, start_depth=0):
@@ -280,95 +245,3 @@ class MergeAssociatesTransformer(NestedTransformer):
         # that moved associations get the correct defining scope
         o._derive_local_symbol_types(parent_scope=o.parent)
         return o
-
-
-def check_if_scalar_syntax(arg, dummy):
-    """
-    Check if an array argument, arg,
-    is passed to an array dummy argument, dummy,
-    using scalar syntax. i.e. arg(1,1) -> d(m,n)
-
-    Parameters
-    ----------
-    arg:   variable
-    dummy: variable
-    """
-    if isinstance(arg, Array) and isinstance(dummy, Array):
-        if arg.dimensions:
-            if not any(isinstance(d, RangeIndex) for d in arg.dimensions):
-                return True
-    return False
-
-
-def transform_sequence_association(routine):
-    """
-    Housekeeping routine to replace scalar syntax when passing arrays as arguments
-    For example, a call like
-
-    .. code-block::
-
-        real :: a(m,n)
-
-        call myroutine(a(i,j))
-
-    where myroutine looks like
-
-    .. code-block::
-
-        subroutine myroutine(a)
-            real :: a(5)
-        end subroutine myroutine
-
-    should be changed to
-
-    .. code-block::
-
-        call myroutine(a(i:m,j)
-
-    Parameters
-    ----------
-    routine : :any:`Subroutine`
-        The subroutine where calls will be changed
-
-    """
-
-    #List calls in routine, but make sure we have the called routine definition
-    calls = (c for c in FindNodes(ir.CallStatement).visit(routine.body) if not c.procedure_type is BasicType.DEFERRED)
-    call_map = {}
-
-    # Check all calls and record changes to `call_map` if necessary.
-    for call in calls:
-        transform_sequence_association_append_map(call_map, call)
-
-    # Fix sequence association in all calls in one go.
-    if call_map:
-        routine.body = Transformer(call_map).visit(routine.body)
-
-def transform_sequence_association_append_map(call_map, call):
-    """
-    Check if `call` contains the sequence association pattern in one of the arguments,
-    and if so, add the necessary transform data to `call_map`.
-    """
-    new_args = []
-    found_scalar = False
-    for dummy, arg in call.arg_map.items():
-        if check_if_scalar_syntax(arg, dummy):
-            found_scalar = True
-
-            n_dims = len(dummy.shape)
-            new_dims = []
-            for s, lower in zip(arg.shape[:n_dims], arg.dimensions[:n_dims]):
-
-                if isinstance(s, RangeIndex):
-                    new_dims += [RangeIndex((lower, s.stop))]
-                else:
-                    new_dims += [RangeIndex((lower, s))]
-
-            if len(arg.dimensions) > n_dims:
-                new_dims += arg.dimensions[len(dummy.shape):]
-            new_args += [arg.clone(dimensions=as_tuple(new_dims)),]
-        else:
-            new_args += [arg,]
-
-    if found_scalar:
-        call_map[call] = call.clone(arguments = as_tuple(new_args))
